requirements.txt
Django==1.5.1
South==0.7.6
django-social-auth==0.7.22
factory-boy==2.0.2
httplib2==0.8
oauth2==1.5.211
py==1.4.13
pytest==2.3.4
pytest-django==2.3.0
python-openid==2.2.5
wsgiref==0.1.2
